fix bug

+++ financial/models/payment.py
@@ -42,7 +42,7 @@ class Payment(models.Model):
 
     status = get_status_field()
 
-    ref_id = models.PositiveIntegerField(null=True, blank=True)
+    ref_id = models.PositiveBigIntegerField(null=True, blank=True)
     ref_status = models.SmallIntegerField(null=True, blank=True)
 
     def create_trx(self):
